Refactor CSS styles for consistency and improve button hover effects
