+++ src/app/Competitions-Dashboard/CompetitionDashboard.module.css
@@ -12,7 +12,7 @@
     display: flex;
     flex-direction: column;
     padding: 1rem;
-    background-color: #333;
+    background-color: #4a4a4a;
     border-radius: 16px;
     min-width: 250px;
     flex: 1;
