+++ src/app/Competitions-Dashboard/EditCompDialog.tsx
@@ -241,13 +241,10 @@ const EditCompDialog = ({
     const [selectedEvents, setSelectedEvents] = useState<
         Record<EventName, EventState>
     >(() => {
-        const initialEvents = EVENTS.reduce(
-            (acc, event) => {
-                acc[event] = { selected: false, rounds: 1 };
-                return acc;
-            },
-            {} as Record<EventName, EventState>,
-        );
+        const initialEvents = EVENTS.reduce((acc, event) => {
+            acc[event] = { selected: false, rounds: 1 };
+            return acc;
+        }, {} as Record<EventName, EventState>);
 
         competition.events.forEach((event) => {
             if (event.name in initialEvents) {
