+++ src/app/Login/Login.module.css
@@ -29,9 +29,7 @@
     border: 2px solid #ccc;
     border-radius: 4px;
     font-size: 16px;
-    transition:
-        border-color 0.3s,
-        box-shadow 0.3s;
+    transition: border-color 0.3s, box-shadow 0.3s;
 }
 
 .username-input:focus,
@@ -43,30 +41,28 @@
 
 .submit-btn {
     padding: 10px 12px;
-    width: 50vw;
+    width: 100%;
     font-size: 18px;
     color: #fff;
-    background-color: #66afe9; /* Blue button */
+    background-color: var(--default-button-background-color);
     border: none;
     border-radius: 4px;
     cursor: pointer;
-    transition:
-        background-color 0.3s,
-        box-shadow 0.3s,
+    transition: background-color 0.3s, box-shadow 0.3s,
         transform 0.3s var(--exponential);
 }
 
 .submit-btn:hover {
     transform: scale(1.05);
-    background-color: #559dd6;
-    box-shadow: 0 0 8px rgb(102 175 233 / 60%);
+    background-color: var(--default-button-background-color);
+    box-shadow: 0 0 8px var(--default-button-shadow-color);
 }
 
 .submit-btn:focus {
     outline: none;
 }
 
-#message {
+.message {
     margin-top: 10px;
     text-align: center;
     color: rgb(255 255 255);
