+++ src/app/Posts/Posts.module.css
@@ -10,17 +10,13 @@
     padding: 10px;
     border-radius: 4px;
     font-size: 16px;
-    transition:
-        border-color 0.3s,
-        box-shadow 0.3s;
+    transition: border-color 0.3s, box-shadow 0.3s;
 }
 
 .title {
     width: 50vw;
     height: 5vh;
-    transition:
-        height 0.3s,
-        width 0.3s var(--exponential);
+    transition: height 0.3s, width 0.3s var(--exponential);
 }
 
 .title:focus {
@@ -34,9 +30,7 @@
     max-height: 30vh;
     overflow-y: auto; /* Enable scrolling when max height is reached */
     height: auto; /* Allows the height to adjust based on content */
-    transition:
-        min-height 0.3s,
-        width 0.3s var(--exponential);
+    transition: min-height 0.3s, width 0.3s var(--exponential);
     transition-delay: 0.2s;
 }
 
@@ -46,7 +40,7 @@
 
 .title:focus,
 .description:focus {
-    border: 2px solid #66afe9;
+    border: 2px solid var(--default-button-background-color);
     box-shadow: 0 0 8px rgb(102 175 233 / 60%);
     outline: none;
 }
@@ -56,7 +50,7 @@
     width: 50vw;
     font-size: 18px;
     color: #fff;
-    background-color: #66afe9; /* Blue button */
+    background-color: var(--default-button-background-color); /* Blue button */
     border: none;
     border-radius: 4px;
     cursor: pointer;
@@ -64,10 +58,7 @@
     justify-content: center;
     align-items: center;
     height: 45px;
-    transition:
-        background-color 0.3s,
-        box-shadow 0.3s,
-        transform 0.3s var(--exponential),
+    transition: background-color 0.3s, box-shadow 0.3s, transform 0.3s,
         height 0.3s;
 }
 
@@ -77,8 +68,8 @@
 
 .post-btn:hover {
     transform: scale(1.05);
-    background-color: #559dd6;
-    box-shadow: 0 0 8px rgb(102 175 233 / 60%);
+    background-color: var(--default-button-hover-background-color);
+    box-shadow: 0 0 8px var(--default-button-hover-background-color) / 60%;
 }
 
 .post-btn:focus {
@@ -138,9 +129,7 @@
     background-color: white;
     border-radius: 2px;
     border: none;
-    transition:
-        border-radius 0.3s var(--exponential),
-        background-color 0.3s;
+    transition: border-radius 0.3s var(--exponential), background-color 0.3s;
 }
 
 .delete-post-btn:hover,
@@ -171,9 +160,7 @@
 }
 
 .edit-title-input {
-    transition:
-        font-size 0.3s,
-        font-weight 0.3s;
+    transition: font-size 0.3s, font-weight 0.3s;
 }
 
 .edit-title-input,
